http://sourceware.org/bugzilla/show_bug.cgi?id=4172
This adds an option to configure that allows --enable-werror and
--disable-werror.  It is enabled by default (backwards compatible).
When --disable-werror is specified as argument to autogen.sh (which
passes it to configure), the -Werror option will *not* be specified
as part of CXXFLAGS, CFLAGS, and GCJFLAGS.  In addition, the hack to
work around the gcj problem (not actually reporting failure in the
event of warnings when -Werror is specified) will be disabled with
--disable-werror is specified.

The implementation of this option has been discussed at great length
on the frysk IRC channel on Mar 8th, 2007.

This commit modifies the earlier move of the "-Wall -Werror" options
from Makefile.rules to frysk-common.ac, by taking out the "-Werror"
option because that one is now under autoconf control.