Push image mapping management into UnwindH.hxx.
Fyi, there's a pre-existing memory-free race bug here; this patch
should help to better understand it.  Prior to this patch the sequence
was:

- frysk calls libunwind to do a step (a.k.a. unwind a frame)

- libunwind's step code calls frysk to populate a proc-info describing
  how to unwind the current frame

- frysk finds and then mmaps in the required image

- frysk creates an ElfImage object from the mmap; during garbage
  collection ElfImage will un-map the mmapped region.

- frysk passes the mmaped region to libunwind's dwarf code which 1)
  fills in the proc-info and 2) retains a reference to the mmaped
  image

- frysk returns to libunwind's step code; this drops all references to
  the ElfImage

- libunwind's dwarf step code uses the supplied unwind information
  including the mmaped region completing the step

- libunwind's dwarf step code "releases" the proc-info locally (frysk
  doesn't know that the mmapped region is unused), and then returns
  control to frysk

Notice how the ElfImage object is dropped before libunwind has
finished with it - a Syste.gc() will destroy the ElfImage and hence
the mmapped region before libunwind has finished with it; and that
frysk doesn't know exactly when it has been finished with.

This patch, consolidates some methods into UnwindH and in the process
makes the race more obvious by blatently creating and then dropping
the ElfImage object.

frysk-sys/lib/unwind/ChangeLog
2008-05-24  Andrew Cagney  <cagney@redhat.com>

 	* Unwind.java (createElfImageFromVDSO): Delete.
 	(createProcInfoFromElfImage): Delete.
 	(fillProcInfoFromVDSO): New.
 	(fillProcInfoFromElfImage): New.
 	* cni/UnwindH.hxx: Implement.
 	* ProcInfo.java: Use.
 	* ElfImage.java: Simplify.
 	* cni/ElfImage.cxx (ElfImage::mapElfImage): Delete.