Provide parent for user and internal exceptions; try.
This adds two root exceptions (they both extend RuntimeException):

UserException:

(I toyed with the name ExternalException also, and better suggestions
welcome).  This corresponds to an error problem caused by user
provided information and hence is likely going to be displayed as an
informational message to the user.  For instance, the user provided
invalid input from the console; or a user supplied file is in some way
invalid, or even a data being read from the target's memory (such as a
table) is found to be invalid.  The key thing is that something
exceptional has occured that is external to frysk's control.

InternalException:

This corresponds to a problem in frysk's code base.  For instance,
missing or unimplemented code, or a programming error.  Java
exceptions such as ArrayIndexOutOfBounds can also be considered to
fall into this category.  If such an exception bubbles up to user
level, it will likely want to be displayed as a backtrace.

frysk-core/frysk/debuginfo/ChangeLog
2008-06-12  Andrew Cagney  <cagney@redhat.com>

	* ObjectDeclarationNotFoundException.java: Extend
	UserException.
	* PieceLocation.java: Throw UserException.

frysk-core/frysk/expr/ChangeLog
2008-06-12  Andrew Cagney  <cagney@redhat.com>

	* ScratchSymTab.java: Throw UserException.

frysk-core/frysk/hpd/ChangeLog
2008-06-12  Andrew Cagney  <cagney@redhat.com>

	* CLI.java (nasty()): Delete.
	(printError(Exception)): Check for UserException.
	* InvalidCommandException.java: Extend UserException.

frysk-core/frysk/util/ChangeLog
2008-06-12  Andrew Cagney  <cagney@redhat.com>

	* CommandlineParser.java: Catch UserException, not
	RuntimeException.

frysk-sys/frysk/ChangeLog
2008-06-12  Andrew Cagney  <cagney@redhat.com>

	* ExternalException.java: New.

frysk-sys/frysk/sys/ChangeLog
2008-06-12  Andrew Cagney  <cagney@redhat.com>

	* Errno.java-sh: Extend ExternalException.
	* cni/Errno.cxx: Update.

frysk-sys/lib/dwfl/ChangeLog
2008-06-12  Andrew Cagney  <cagney@redhat.com>

	* ElfException.java: Extend UserException.